User functions split into 3 classes. UserManager, UserRegistrationManager and PasswordManager.
